Gnombl za zolotom!

+++ account_app/src/main/java/com/example/account_app/service/user/UserService.java
@@ -2,6 +2,7 @@ package com.example.account_app.service.user;
 
 import com.example.account_app.model.User;
 import com.example.account_app.repository.UserRepository;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -9,11 +10,12 @@ import java.util.Optional;
 
 @Service
 public class UserService {
-
+    private final PasswordEncoder passwordEncoder;
     private final UserRepository userRepository;
 
-    public UserService(UserRepository userRepository) {
+    public UserService(UserRepository userRepository, PasswordEncoder passwordEncoder) {
         this.userRepository = userRepository;
+        this.passwordEncoder = passwordEncoder;
     }
 
     public List<User> getAllUsers() {
@@ -25,10 +27,23 @@ public class UserService {
     }
 
     public User saveUser(User user) {
+        // Хэшируем пароль перед сохранением
+        user.setPassword(passwordEncoder.encode(user.getPassword()));
         return userRepository.save(user);
     }
 
+    public boolean existsByLogin(String login) {
+        return userRepository.existsByLogin(login);
+    }
+
     public void deleteUser(Integer id) {
+        if (!userRepository.existsById(id)) {
+            throw new RuntimeException("User not found with id: " + id);
+        }
         userRepository.deleteById(id);
+    }
+
+    public Optional<User> getUserByLogin(String login) {
+        return userRepository.findByLogin(login);
     }
 }
